Created the Main Screen

+++ app/src/main/java/com/who/climasense/models/Weather.kt
@@ -5,5 +5,5 @@ data class Weather(
     val cnt: Int,
     val cod: String,
     val list: List<WeatherItem>,
-    val message: Int
+    val message: Double
 )
